server/requirements.txt
asgiref==3.5.0
astroid==2.11.3
beautifulsoup4==4.11.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
dill==0.3.4
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pylint==2.13.7
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
whitenoise==6.1.0
wrapt==1.14.0
